+++ lib/ex_airtable/airtable/record.ex
@@ -50,22 +50,32 @@ defmodule ExAirtable.Airtable.Record do
 
   ## Examples
 
-      iex> record = %ExAirtable.Airtable.Record{fields: %{"AirtableField" => "value"}}
+      iex> record = %ExAirtable.Airtable.Record{id: "1", fields: %{"AirtableField" => "value"}}
 
       iex> to_schema(record, %{"AirtableField" => "localfield"})
-      %{"localfield" => "value"}
+      %{"airtable_id" => "1", "localfield" => "value"}
 
       iex> to_schema(record, nil)
-      %{"AirtableField" => "value"}
+      %{"airtable_id" => "1", "AirtableField" => "value"}
   """
-  def to_schema(%__MODULE__{} = record, nil), do: record.fields
+  def to_schema(%__MODULE__{} = record, nil) do 
+    Map.merge(
+      %{"airtable_id" => record.id},
+      record.fields
+    )
+  end
 
   def to_schema(%__MODULE__{} = record, schema_map) when is_map(schema_map) do
-    Enum.reduce(record.fields, %{}, fn {key, val}, acc ->
+    schema = Enum.reduce(record.fields, %{}, fn {key, val}, acc ->
       case Map.get(schema_map, key) do
         nil -> acc
         new_key -> Map.put(acc, new_key, val)
       end
     end)
+
+    case is_binary(Enum.at(Map.keys(schema), 0)) do
+      true -> Map.merge(schema, %{"airtable_id" => record.id})
+      false -> Map.merge(schema, %{airtable_id: record.id})
+    end
   end
 end