+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # CHANGELOG
 
+- 0.2.6
+  - Add Airtable ID to exported table schemas. See `ExAirtable.Table.to_schema` and `ExAirtable.Airtable.Record.to_schema` for more details.
 - 0.2.5
   - Make `Table.list_async` overrideable in situations where you want to change the default table listing parameters.
   - Clean up `params` for `Table.list` and `Table.list_async` to use Keyword lists in both places (instead of a map in one and a keyword list in the other)